+++ frontend/src/redux/selectors/filtersSelectors.ts
@@ -1,3 +1,0 @@
-import type { RootState } from "redux/store";
-
-export const selectMenuFilters = (state: RootState) => state.filters.menu;